Use DaemonSet + AWS NLB + Session Stickiness
Due to architectural restrictions in AWS NLB, it's not possible to
enable stickiness when using TLS encryption between the LB and the
targets. Customers that have faced such requirements have been
instructed by Amazon support to move TLS encryption back in the chain,
onto the app servers directly. The traffic can then be passed through
the NLB as TCP traffic and not TLS traffic, and session stickiness is
enabled on the NLB directly, without compromising complete end-to-end
encryption in the environment.

See https://confluence.atlassian.com/jirakb/sticky-sessions-for-aws-nlb-over-tls-1064091930.html